- style: add filter styles - feat: filter placeholder - refactor: move props up to overview panel component - feat: show dropdown for sample change
