+++ src/css/VisOverviewPanel.css
@@ -90,85 +90,91 @@
             display: flex;
             flex-direction: row;
             align-items: center;
+            justify-content: space-between;
             width: 100%;
-            height: 100px;
             border-right: 1px solid lightgrey;
             background-color: #f3f9ff;
             border-bottom: 1px solid #aabacd;
-            padding: 20px;
+            padding: 25px 20px;
 
-            .cancer-selector {
+            .dropdown-container {
+                position: relative;
                 display: flex;
-                flex-direction: column;
-                padding: 0px;
+                width: max-content;
 
-                .dropdown-container {
-                    position: relative;
+                .dropdown-button {
                     display: flex;
+                    align-items: center;
+                    font-size: 1.875rem;
+                    font-weight: 600;
+                    background-color: transparent;
+                    border: none;
+                    border-radius: 5px;
+                    padding: 0px 15px;
+
+                    .icon {
+                        transform: rotate(180deg);
+                        /* transition: transform 0.2s ease; */
+                        width: 20px;
+                        margin-left: 20px;
+                    }
+                }
+
+                .dropdown-button:hover {
+                    cursor: pointer;
+                    background-color: #dee5ec;
+                }
+
+                .dropdown-button.toggle-open {
+                    .icon {
+                        transition: transform 0.2s ease;
+                        transform: rotate(0deg);
+                    }
+                }
+
+                .dropdown-items {
+                    position: absolute;
+                    display: flex;
+                    flex-direction: column;
+                    background-color: white;
+                    border: 1px solid #889cb5;
+                    border-radius: 5px;
+                    padding: 10px;
+                    height: auto;
                     width: max-content;
+                    font-size: 1rem;
+                    overflow: auto scroll;
+                    z-index: 1;
+                    top: 100%;
+                    gap: 5px;
+                }
 
-                    .dropdown-button {
-                        display: flex;
-                        align-items: center;
-                        font-size: 1.875rem;
-                        font-weight: 600;
-                        background-color: transparent;
-                        border: none;
-                        border-radius: 5px;
-                        padding: 0px 15px;
+                /* .dropdown-items.d-none {
+                    opacity: 0;
+                    height: 10px;
+                    transition: height 2s ease;
+                }
 
-                        .icon {
-                            transform: rotate(180deg);
-                            /* transition: transform 0.2s ease; */
-                            width: 20px;
-                            margin-left: 20px;
-                        }
-                    }
+                .dropdown-items.d-flex {
+                    opacity: 1;
+                    height: 100px;
+                    transition: height 2s ease;
+                } */
 
-                    .dropdown-button:hover {
-                        cursor: pointer;
-                        background-color: #dee5ec;
-                    }
+                .dropdown-items button {
+                    display: flex;
+                    justify-content: space-between;
+                    padding: 5px 15px;
+                    border-radius: 5px;
+                    font-size: 1.25rem;
+                    background-color: transparent;
+                    border: none;
+                    gap: 30px;
+                }
 
-                    .dropdown-button.toggle-open {
-                        .icon {
-                            transition: transform 0.2s ease;
-                            transform: rotate(0deg);
-                        }
-                    }
-
-                    .dropdown-items {
-                        position: absolute;
-                        display: flex;
-                        flex-direction: column;
-                        background-color: white;
-                        border: 1px solid #889cb5;
-                        border-radius: 5px;
-                        padding: 10px;
-                        height: 500px;
-                        font-size: 1rem;
-                        overflow: auto scroll;
-                        z-index: 1;
-                        top: 100%;
-                        gap: 5px;
-                        width: max-content;
-                    }
-
-                    .dropdown-items button {
-                        display: flex;
-                        justify-content: space-between;
-                        padding: 5px 15px;
-                        border-radius: 5px;
-                        font-size: 1.25rem;
-                        background-color: transparent;
-                        border: none;
-                        gap: 30px;
-                    }
-
-                    .dropdown-items button:hover {
-                        cursor: pointer;
-                        background-color: #c6def6;
-                    }
+                .dropdown-items button:hover {
+                    cursor: pointer;
+                    background-color: #c6def6;
                 }
             }
 
@@ -200,6 +206,37 @@
                         font-weight: 600;
                         margin-left: 10px;
                     }
+                }
+            }
+
+            .cancer-selector {
+                display: flex;
+                flex-direction: column;
+                padding: 0px;
+            }
+        }
+
+        .overview-controls {
+            padding: 10px 20px;
+            display: flex;
+            gap: 15px;
+            flex-direction: row;
+            flex-wrap: wrap;
+            border-top: 1px solid #d3d3d3;
+            border-bottom: 1px solid #d3d3d3;
+
+            .filter {
+                width: 210px;
+
+                select.dropdown {
+                    font-weight: 500;
+                    padding: 5px 10px;
+                    color: #617c9d;
+                    border: 1px solid #617c9d;
+                    border-radius: 4px;
+                    max-width: 250px;
+                    width: 100%;
+                    text-overflow: ellipsis;
                 }
             }
         }
